Final/requirements.txt
pyserial
flask
opencv-python
numpy
